- show aps-databases also customer-mysql overview, fixes #272 - aps-databases are now increasing the customers mysql-contingent, refs #278 - customers viewing the aps-list are now notified if a package requires a database but the contingent is used up - the update will add all web[X]aps[Y] databases to the according customer regardless of contingent - fix tiny issue in preconfig - setting version to 0.9.10-svn2
refs #272,278

+++ lib/functions/froxlor/function.openRootDB.php
@@ -26,7 +26,7 @@
  *
  * @return	null
  */
-function openRootDB($debugHandler, $lockfile)
+function openRootDB($debugHandler = false, $lockfile = false)
 {
 	global $db_root;
 
@@ -47,14 +47,22 @@ function openRootDB($debugHandler, $lockfile)
 		/**
 		 * Do not proceed further if no database connection could be established
 		 */
-
-		fclose($debugHandler);
-		unlink($lockfile);
+		if(isset($debugHandler) && $debugHandler !== false)
+		{
+			fclose($debugHandler);
+		}
+		if(isset($lockfile) && $lockfile !== false) 
+		{
+			unlink($lockfile);	
+		}
 		die('root can\'t connect to mysqlserver. Please check userdata.inc.php! Exiting...');
 	}
 
 	unset($db_root->password);
-	fwrite($debugHandler, 'Database-rootconnection established' . "\n");
+	if(isset($debugHandler) && $debugHandler !== false)
+	{
+		fwrite($debugHandler, 'Database-rootconnection established' . "\n");
+	}
 	
 	unset($sql);
 }
